+++ test/unit/string/index.ts
@@ -1,0 +1,7 @@
+import { injectIntegerTests } from "./integer.js";
+
+export function injectStringTests() {
+  describe("string", () => {
+    injectIntegerTests();
+  });
+}